Bug #198680: Added method to sort files in zip
This allows packaging code to sort files in the zip when
doing a release package. Sorting files reduces the
number of changes between APKs.

+++ build-system/builder/src/main/java/com/android/builder/internal/packaging/zip/FileUseMap.java
@@ -57,7 +57,7 @@ class FileUseMap {
      *
      * @param size the size of the file
      */
-    FileUseMap(int size) {
+    FileUseMap(long size) {
         Preconditions.checkArgument(size >= 0, "size < 0");
 
         mSize = size;
